Generalize 'broadcast' use-case for 'servant' and 'yesod'
