+++ front.cabal
@@ -30,11 +30,17 @@ library
                      , Text.Blaze.Front.Event
                      , Text.Blaze.Front.Renderer
                      , Web.Front
+                     , Web.Front.Broadcast
                      -- shared
                      , Bridge
   -- other-modules:
   other-extensions:    FlexibleInstances, TypeSynonymInstances, OverloadedStrings, GeneralizedNewtypeDeriving, Rank2Types, ExistentialQuantification, DeriveDataTypeable, MultiParamTypeClasses, DeriveFunctor, FunctionalDependencies, ScopedTypeVariables
   build-depends:       base <4.12
+                     , aeson
+                     , async
+                     , stm
+                     , stm-lifted
+                     , conduit
                      , text >=1.2 && <1.3
                      , mtl
                      , bytestring >=0.10 && <0.11
@@ -44,6 +50,7 @@ library
                      , fay-dom
                      , fay-websockets
                      , websockets
+  ghc-options:    -main-is YesodTodo -Wall
   hs-source-dirs:      src, shared
   default-language:    Haskell2010
 
@@ -86,5 +93,27 @@ executable todo-servant-example
                , wai            >= 3.0 
                , warp           >= 3.0
                , front
-  other-modules: Shared
-  ghc-options:    -main-is ServantTodo
+  other-modules: Shared, Todo
+  ghc-options:    -main-is ServantTodo -Wall
+  default-language: Haskell2010
+
+executable todo-yesod-example
+  main-is:             YesodTodo.hs
+  hs-source-dirs:      examples/todo
+  build-depends: base           >= 4.7 && <4.12
+               , front
+               , text
+               , mtl
+               , fay
+               , random
+               , yesod-core
+               , yesod-static
+               , yesod-websockets
+               , conduit
+               , aeson
+               , stm-lifted
+               , bytestring
+
+  other-modules: Shared, Todo
+  ghc-options:    -main-is YesodTodo -Wall
+  default-language: Haskell2010
